Give “hello” page its own folder
